vessence/vault_web/requirements.txt
fastapi==0.135.1
starlette>=0.46.0,<1.0.0
uvicorn[standard]==0.41.0
jinja2==3.1.6
python-multipart==0.0.22
aiofiles==23.2.1
httpx==0.28.1
pillow==11.3.0
requests==2.32.5
chromadb==1.5.4
python-dotenv==1.2.2
itsdangerous==2.2.0
authlib==1.6.9
pyotp==2.9.0
litellm>=1.82.0
tiktoken>=0.12.0
